Fixed API configuration
